fix: change share mechanism
