+++ react-native-src/navigation/AppDrawerNavigation.tsx
@@ -1,17 +1,25 @@
 import React from 'react';
-import {createDrawerNavigator} from '@react-navigation/drawer';
+import {
+  createDrawerNavigator,
+  DrawerContentScrollView,
+  DrawerItemList,
+  DrawerItem,
+  DrawerContentComponentProps,
+} from '@react-navigation/drawer';
 import {MainStackNavigation} from './MainStackNavigation';
 import {AuthorizationFlowStackNavigation} from './AuthorizationFlowStackNavigation';
-import {
-  ShareScreen,
-  ShareScreenNavigationHeader,
-  drawerShareIcon,
-} from '../pages/ShareScreen/ShareScreen';
+import {drawerShareIcon, onShare} from './ShareScreen';
 
+const CustomDrawerContent: React.FC<DrawerContentComponentProps> = props => (
+  <DrawerContentScrollView {...props}>
+    <DrawerItemList {...props} />
+    <DrawerItem label="Share" icon={drawerShareIcon} onPress={onShare} />
+  </DrawerContentScrollView>
+);
 const Drawer = createDrawerNavigator();
 
 export const AppDrawerNavigation = () => (
-  <Drawer.Navigator>
+  <Drawer.Navigator drawerContent={CustomDrawerContent}>
     <Drawer.Screen
       name="Home"
       component={MainStackNavigation}
@@ -24,15 +32,6 @@ export const AppDrawerNavigation = () => (
         headerShown: false,
       }}
       component={AuthorizationFlowStackNavigation}
-    />
-    <Drawer.Screen
-      name="Share"
-      options={{
-        title: 'Share link to store',
-        header: props => <ShareScreenNavigationHeader {...props} />,
-        drawerIcon: drawerShareIcon,
-      }}
-      component={ShareScreen}
     />
   </Drawer.Navigator>
 );
